refactor: change quantity to double from int

+++ src/main/java/com/group7/krisefikser/dto/response/AggregatedStorageItemResponse.java
@@ -17,6 +17,6 @@ import lombok.NoArgsConstructor;
 public class AggregatedStorageItemResponse {
   private int itemId;
   private ItemResponse item;
-  private int totalQuantity;
+  private double totalQuantity;
   private LocalDateTime earliestExpirationDate;
 }
